requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
httpx==0.28.1
redis==5.2.1
jinja2==3.1.6
python-dotenv==1.0.1
llama-cpp-python==0.3.20
pydantic[email]==2.10.4
pydantic-settings==2.7.0
aioredis==2.0.1
